requirements.txt
fastapi
uvicorn
python-dotenv
pyyaml
elasticsearch==8.12.0
httpx
mcp
aiohttp